refactor(repositories): co-locate modules
