+++ src/repositories/athenai-repository/__tests__/athenai-repository-v2.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
-import { AthenaiRepositoryV2, mergeSourcesV2 } from '../athenai-repository-v2';
-import { getEffectiveHeadsign } from '../../domain/transit/get-effective-headsign';
+import { AthenaiRepositoryV2 } from '..';
+import { getEffectiveHeadsign } from '../../../domain/transit/get-effective-headsign';
 import {
   TestDataSourceV2,
   createFixtureV2,
@@ -13,349 +13,16 @@ import {
   AFTER_BOUNDARY_PAST,
   EXCEPTION_HOLIDAY,
 } from './fixtures/test-data-source-v2';
-import { minutesToDate } from '../../domain/transit/calendar-utils';
-import type { SourceDataV2 } from '../../datasources/transit-data-source-v2';
+import { minutesToDate } from '../../../domain/transit/calendar-utils';
 
-/** Assert result is successful and return narrowed type for safe data access. */
-function assertSuccess<T>(result: {
-  success: boolean;
-}): asserts result is { success: true; data: T } {
+/** Assert result is successful and preserve the successful result shape. */
+function assertSuccess<T extends { success: boolean }>(
+  result: T,
+): asserts result is Extract<T, { success: true }> {
   if (!result.success) {
     throw new Error(`Expected success but got failure: ${JSON.stringify(result)}`);
   }
 }
-
-// ---------------------------------------------------------------------------
-// mergeSourcesV2
-// ---------------------------------------------------------------------------
-
-describe('mergeSourcesV2', () => {
-  it('converts stops with empty agency_id', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    // Use l=0 stop (l=1 stops are filtered out)
-    const tdn01 = merged.stops.find((s) => s.stop_id === 'tdn_01');
-    expect(tdn01).toBeDefined();
-    expect(tdn01!.agency_id).toBe('');
-  });
-
-  it('resolves stop_names from translations', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    const tdn01 = merged.stops.find((s) => s.stop_id === 'tdn_01');
-    expect(tdn01!.stop_names).toEqual({ ja: '新庚申塚', en: 'Shin-koshinzuka' });
-  });
-
-  it('excludes location_type=1 stops (parent stations filtered until UI supports grouping)', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    const station = merged.stops.find((s) => s.stop_id === 'sta_parent');
-    expect(station).toBeUndefined();
-    // All stops should be l=0
-    for (const s of merged.stops) {
-      expect(s.location_type).toBe(0);
-    }
-  });
-
-  it('builds routeMap from v2 routes', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    expect(merged.routeMap.size).toBe(5);
-    const subway = merged.routeMap.get('route_subway');
-    expect(subway).toBeDefined();
-    expect(subway!.route_type).toBe(1);
-    expect(subway!.agency_id).toBe('test:agency');
-  });
-
-  it('builds agencyMap with translations', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    expect(merged.agencyMap.size).toBe(2);
-    const agency = merged.agencyMap.get('test:agency');
-    expect(agency!.agency_names).toEqual({ ja: 'テスト事業者', en: 'Test Agency' });
-  });
-
-  it('builds resolvedPatterns from tripPatterns', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    // 10 original patterns + 1 circular (tp_bus_c)
-    expect(merged.resolvedPatterns.size).toBe(13);
-    const subN = merged.resolvedPatterns.get('tp_sub_n');
-    expect(subN).toBeDefined();
-    expect(subN!.route.route_id).toBe('route_subway');
-    expect(subN!.headsign).toBe('Nishi-takashimadaira');
-  });
-
-  it('builds stopRouteTypeMap via tripPattern FK', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    expect(merged.stopRouteTypeMap.get('sub_02')).toEqual([1, 3]);
-    expect(merged.stopRouteTypeMap.get('tdn_04')).toEqual([0, 2]);
-  });
-
-  it('builds stopsMetaMap with agencies via tripPattern -> route -> agency', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    const meta = merged.stopsMetaMap.get('bus_01');
-    expect(meta).toBeDefined();
-    const ids = meta!.agencies.map((a) => a.agency_id).sort();
-    expect(ids).toEqual(['test:agency', 'test:partner']);
-  });
-
-  it('builds sourceMeta from feedInfo', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    expect(merged.sourceMetas).toHaveLength(1);
-    expect(merged.sourceMetas[0].id).toBe('test');
-    expect(merged.sourceMetas[0].validity.startDate).toBe('20260101');
-  });
-
-  it('merges lookup data', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-    expect(merged.lookup).toBeDefined();
-  });
-
-  it('converts v2 optional stop fields (wb, ps, pc)', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-
-    // bus_01 has wb=1, ps='sta_parent', pc='1'
-    const bus01 = merged.stops.find((s) => s.stop_id === 'bus_01');
-    expect(bus01).toBeDefined();
-    expect(bus01!.wheelchair_boarding).toBe(1);
-    expect(bus01!.parent_station).toBe('sta_parent');
-    expect(bus01!.platform_code).toBe('1');
-
-    // bus_03 has wb=2, no ps/pc
-    const bus03 = merged.stops.find((s) => s.stop_id === 'bus_03');
-    expect(bus03).toBeDefined();
-    expect(bus03!.wheelchair_boarding).toBe(2);
-    expect(bus03!.parent_station).toBeUndefined();
-    expect(bus03!.platform_code).toBeUndefined();
-
-    // tdn_01 has no v2 optional fields
-    const tdn01 = merged.stops.find((s) => s.stop_id === 'tdn_01');
-    expect(tdn01).toBeDefined();
-    expect(tdn01!.wheelchair_boarding).toBeUndefined();
-    expect(tdn01!.parent_station).toBeUndefined();
-    expect(tdn01!.platform_code).toBeUndefined();
-  });
-
-  it('converts TripPatternJson to app-internal TripPattern with object array stops', () => {
-    const fixture = createFixtureV2();
-    const merged = mergeSourcesV2([fixture]);
-
-    // tripPatterns should be converted from JSON schema to app-internal type
-    const pattern = merged.tripPatterns.get('tp_sub_n');
-    expect(pattern).toBeDefined();
-    expect(pattern!.route_id).toBe('route_subway');
-    expect(pattern!.headsign).toBe('Nishi-takashimadaira');
-    // stops must be object array with id (and optional headsign/shapeDistTraveled)
-    expect(pattern!.stops).toEqual([{ id: 'sub_01' }, { id: 'sub_02' }, { id: 'sub_03' }]);
-
-    // direction=0 must be preserved (not dropped by falsy check)
-    expect(pattern!.direction).toBe(0);
-
-    // direction=1 must also be preserved
-    const patternM = merged.tripPatterns.get('tp_sub_m');
-    expect(patternM).toBeDefined();
-    expect(patternM!.direction).toBe(1);
-
-    // direction should be undefined when dir is not set in JSON
-    const busPattern = merged.tripPatterns.get('tp_bus_o');
-    expect(busPattern).toBeDefined();
-    expect(busPattern!.direction).toBeUndefined();
-
-    // empty headsign should be preserved
-    const emptyH = merged.tripPatterns.get('tp_ptr_e');
-    expect(emptyH).toBeDefined();
-    expect(emptyH!.headsign).toBe('');
-    expect(emptyH!.stops).toEqual([{ id: 'bus_01' }]);
-  });
-
-  it('preserves stop_headsign (sh) in TripPattern.stops', () => {
-    const merged = mergeSourcesV2([createFixtureV2()]);
-    const pattern = merged.tripPatterns.get('tp_ptr_sh');
-    expect(pattern).toBeDefined();
-    expect(pattern!.headsign).toBe('');
-    expect(pattern!.stops).toEqual([
-      { id: 'bus_01', headsign: 'Oji-eki via Park' },
-      { id: 'bus_02', headsign: 'Oji-eki' },
-    ]);
-  });
-
-  it('preserves stop_headsign alongside non-empty trip_headsign', () => {
-    const merged = mergeSourcesV2([createFixtureV2()]);
-    const pattern = merged.tripPatterns.get('tp_bus_sh');
-    expect(pattern).toBeDefined();
-    expect(pattern!.headsign).toBe('Oji-eki via All Stops');
-    expect(pattern!.stops[0]).toEqual({ id: 'bus_01', headsign: 'Oji-eki via All Stops' });
-    expect(pattern!.stops[1]).toEqual({ id: 'bus_02', headsign: 'Oji-eki' });
-    // bus_03 has no stop_headsign
-    expect(pattern!.stops[2]).toEqual({ id: 'bus_03' });
-  });
-
-  describe('i18n pre-compilation uses feed_lang (not agency_lang)', () => {
-    /**
-     * Minimal fixture where feed_lang="it" but agency_lang="en".
-     * Verifies that base values are injected under feed_lang, not agency_lang.
-     */
-    function createFeedLangFixture(): SourceDataV2 {
-      return {
-        prefix: 'itfeed',
-        data: {
-          bundle_version: 3,
-          kind: 'data',
-          stops: {
-            v: 2,
-            data: [{ v: 2, i: 'itfeed:s1', n: 'Stazione Centrale', a: 45.0, o: 9.0, l: 0 }],
-          },
-          routes: {
-            v: 2,
-            data: [
-              {
-                v: 2,
-                i: 'itfeed:r1',
-                s: 'L1',
-                l: 'Linea 1',
-                t: 3,
-                c: '',
-                tc: '',
-                ai: 'itfeed:ag1',
-              },
-            ],
-          },
-          agency: {
-            v: 2,
-            data: [
-              {
-                v: 2,
-                i: 'itfeed:ag1',
-                n: 'Compagnia Trasporti',
-                u: 'https://example.com',
-                tz: 'Europe/Rome',
-                l: 'en',
-              },
-            ],
-          },
-          calendar: { v: 1, data: { services: [], exceptions: [] } },
-          feedInfo: {
-            v: 1,
-            data: { pn: 'Test', pu: 'https://example.com', l: 'it', s: '', e: '', v: '' },
-          },
-          timetable: { v: 2, data: {} },
-          tripPatterns: {
-            v: 2,
-            data: {
-              tp1: {
-                v: 2,
-                r: 'itfeed:r1',
-                h: 'Stazione Centrale',
-                stops: [{ id: 'itfeed:s1' }],
-              },
-            },
-          },
-          translations: {
-            v: 1,
-            data: {
-              agency_names: {
-                'itfeed:ag1': { en: 'English Transit Co.' },
-              },
-              route_long_names: {
-                'itfeed:r1': { en: 'Line 1' },
-              },
-              route_short_names: {},
-              stop_names: {
-                'itfeed:s1': { en: 'Central Station' },
-              },
-              trip_headsigns: {
-                'Stazione Centrale': { en: 'Central Station' },
-              },
-              stop_headsigns: {},
-            },
-          },
-          lookup: { v: 2, data: {} },
-        },
-      };
-    }
-
-    it('injects headsign base value under feed_lang, not agency_lang', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const translations = merged.headsignTranslations.get('itfeed');
-      // feed_lang="it", so "Stazione Centrale" should be injected as "it" candidate
-      expect(translations!.trip_headsigns['Stazione Centrale']).toEqual({
-        it: 'Stazione Centrale',
-        en: 'Central Station',
-      });
-      // NOT agency_lang="en" — en already exists from translations.txt
-    });
-
-    it('injects stop_names base value under feed_lang', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const stop = merged.stops.find((s) => s.stop_id === 'itfeed:s1');
-      expect(stop!.stop_names).toEqual({
-        it: 'Stazione Centrale',
-        en: 'Central Station',
-      });
-    });
-
-    it('injects route_long_names base value under feed_lang', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const route = merged.routeMap.get('itfeed:r1');
-      expect(route!.route_long_names).toEqual({
-        it: 'Linea 1',
-        en: 'Line 1',
-      });
-    });
-
-    it('injects route_short_names base value under feed_lang', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const route = merged.routeMap.get('itfeed:r1');
-      expect(route!.route_short_names).toEqual({ it: 'L1' });
-    });
-
-    it('merges route_short_names translations from pipeline into Route', () => {
-      // Simulate a pipeline-emitted route_short_name translation
-      // (e.g. kyoto-city-bus `市バス1` → `1 City Bus`).
-      const source = createFeedLangFixture();
-      source.data.translations.data.route_short_names = {
-        'itfeed:r1': { en: 'L1 Express' },
-      };
-      const merged = mergeSourcesV2([source]);
-      const route = merged.routeMap.get('itfeed:r1');
-      expect(route!.route_short_names).toEqual({
-        it: 'L1',
-        en: 'L1 Express',
-      });
-      // route_long_names is unaffected by the short_names translation
-      expect(route!.route_long_names).toEqual({
-        it: 'Linea 1',
-        en: 'Line 1',
-      });
-    });
-
-    it('injects agency_names base value under feed_lang, not agency_lang', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const agency = merged.agencyMap.get('itfeed:ag1');
-      // agency_name base value should be injected under feed_lang="it".
-      expect(agency!.agency_names).toEqual({
-        en: 'English Transit Co.',
-        it: 'Compagnia Trasporti',
-      });
-    });
-
-    it('uses empty names when agency has no entry in agency-attributes', () => {
-      const merged = mergeSourcesV2([createFeedLangFixture()]);
-      const agency = merged.agencyMap.get('itfeed:ag1');
-      // itfeed:ag1 has no entry in AGENCY_ATTRIBUTES, so display names are empty
-      expect(agency!.agency_short_names).toEqual({});
-      expect(agency!.agency_long_names).toEqual({});
-      expect(agency!.agency_short_name).toBe('');
-      expect(agency!.agency_long_name).toBe('');
-    });
-  });
-});
 
 // ---------------------------------------------------------------------------
 // AthenaiRepositoryV2.create
