+++ src/repositories/local-storage-user-data-repository/index.ts
@@ -1,0 +1,1 @@
+export { LocalStorageUserDataRepository } from './local-storage-user-data-repository';
