+++ src/repositories/mock-repository/index.ts
@@ -1,0 +1,1 @@
+export { MockRepository } from './mock-repository';
